Move old config to deprecated module, make reset db part of sync worker, fail if worker type is not specified
